Enterprise-Decision-Intelligence/requirements.txt
pandas>=2.0
numpy>=1.24
scipy>=1.11
python-dotenv>=1.0
